refactor(cms-sync): extract engine and config, rewrite monolith as thin wrapper
